更改所有主键类型为LANG，并更改订单表关联为管理账户ID非用户ID

+++ zeus-common-core/src/main/java/com/ptb/zeus/common/core/mapper/user/TbRoleMapper.java
@@ -16,5 +16,5 @@ import java.util.List;
 @Component
 public interface TbRoleMapper extends AutoMapper<TbRole> {
 
-	List<TbRole> selectByUID(@Param("id") Integer id);
+	List<TbRole> selectByUID(@Param("id") Long id);
 }
