+++ zeus-main-facade/src/main/java/com/ptb/zeus/service/main/IMUserServiceService.java
@@ -11,5 +11,5 @@ import com.baomidou.framework.service.ISuperService;
 public interface IMUserServiceService extends ISuperService<MUserService> {
 
 
-	void buyProductService(Integer uid, int productID);
+	void buyProductService(Long uid, Long productID);
 }
